Added the new programs and changes to existing codes

+++ GuessingOutput/src/main/java/com/java2blog/output/leetcode/SquareRoot.java
@@ -4,8 +4,8 @@ public class SquareRoot {
 	public static int mySqrt(int x) {
 		long left = 0, right = x;
 		while (left < right) {
-			long mid = left + ((long) right - left + 1) / 2;
-			if (mid * mid <= (long) x)
+			long mid = left + (right - left + 1) / 2;
+			if (mid * mid <=  x)
 				left = mid;
 			else
 				right = mid - 1;
